chore: filter tests passing
